refactor(parser): split primary/var.rs into cohesive submodules (§7-8)
Split the 1329-line parser/primary/var.rs (over the 500-line limit) into a
facade plus five themed submodules under src/parser/primary/var/. All files
are under 500 lines.

var.rs is free parser functions addressed as var::<name> (14 external call
sites in 8 files); the facade re-exports each submodule preserving its
original visibility, so no external caller changes. primary/mod.rs is
untouched.

- var.rs (29)          - facade: mod decls + re-exports
- ident.rs (159)       - parse_ident_with_hyphens, parse_qualified_ident_*,
                         parse_interpolation_*, is_pseudo_package
- adverb.rs (122)      - parse_var_name_adverb_suffixes, parse_adverb_value_pub,
                         adverb bracket helpers (find_matching_bracket, etc.)
- perl5.rs (212)       - detect_perl5_scalar_var, detect_perl5_sigil_var
- scalar.rs (389)      - scalar_var, parse_var_name_from_str,
                         parse_dollar_paren_block_pub, symbolic-deref helpers
- sigil_vars.rs (488)  - array_var, hash_var, code_var,
                         parse_operator_code_ref_suffix

Pure relocation. Moved fns are pub(crate) in submodules and re-exported at
their original visibility; body refs to primary-level siblings rewritten to
absolute crate::parser paths. Intra-var/ cross-module calls use super::
sibling paths. The fn multiset is byte-identical. No behaviour change.

External var:: callers preserved without edits:
  src/parser/expr/postfix/dot_assign.rs   - parse_ident_with_hyphens
  src/parser/expr/postfix/loop_.rs        - parse_ident_with_hyphens (×2)
  src/parser/stmt/class/package_decl.rs   - is_pseudo_package
  src/parser/stmt/idents.rs               - detect_perl5_scalar_var,
                                            parse_adverb_value_pub
  src/parser/primary/misc/colonpair.rs    - parse_ident_with_hyphens
  src/parser/primary/misc/reduction.rs    - parse_ident_with_hyphens
  src/parser/primary/misc/anon_decl.rs   - parse_ident_with_hyphens
  src/parser/primary/regex/lit.rs         - parse_ident_with_hyphens
  src/parser/primary/string/interp_var.rs - parse_var_name_from_str
  src/parser/primary/container/sigil_context.rs - parse_dollar_paren_block_pub
  src/parser/primary/mod.rs              - scalar_var, array_var, hash_var,
                                           code_var
