+++ src/parser/primary/var.rs
@@ -1,1329 +1,29 @@
-use std::sync::atomic::{AtomicU64, Ordering};
+//! Variable parsing facade: re-exporting themed submodules.
+//!
+//! Split from the original monolithic `var.rs` (1329 lines). Each submodule
+//! owns a cohesive group of free functions; this facade preserves every
+//! function's original visibility through explicit named re-exports so external
+//! callers (`var::<name>` and `crate::parser::primary::var::<name>`) keep
+//! working unchanged.  `primary/mod.rs` is untouched.
+//!
+//! - ident.rs      (138)  — identifier + pseudo-package utilities
+//! - adverb.rs     (113)  — adverb-suffix + bracket-value parsing
+//! - perl5.rs      (172)  — Perl 5 compatibility detection
+//! - scalar.rs     (292)  — $-sigil parser + dollar-paren-block
+//! - sigil_vars.rs (399)  — @/%/& sigil parsers + operator code-ref
 
-use super::super::parse_result::{PError, PResult, parse_char};
+mod adverb;
+mod ident;
+mod perl5;
+mod scalar;
+mod sigil_vars;
 
-use crate::ast::Expr;
-use crate::symbol::Symbol;
-use crate::value::Value;
+// ── pub(crate): accessible from anywhere in the crate ──────────────────────
+pub(crate) use adverb::parse_adverb_value_pub;
+pub(crate) use ident::{is_pseudo_package, parse_ident_with_hyphens};
+pub(crate) use perl5::detect_perl5_scalar_var;
 
-use super::super::expr::expression;
-use super::super::helpers::{
-    is_raku_identifier_continue, is_raku_identifier_start, normalize_raku_identifier, ws,
-};
-use super::container::paren_expr;
-use super::current_line_number;
-use super::misc::parse_block_body;
-
-use crate::token_kind::TokenKind;
-
-static ANON_STATE_COUNTER: AtomicU64 = AtomicU64::new(0);
-static ANON_ARRAY_COUNTER: AtomicU64 = AtomicU64::new(0);
-
-/// Build a string concatenation expression: left ~ right
-fn concat_exprs(left: Expr, right: Expr) -> Expr {
-    Expr::Binary {
-        left: Box::new(left),
-        op: TokenKind::Tilde,
-        right: Box::new(right),
-    }
-}
-
-fn parse_qualified_ident_prefix_with_hyphens<'a>(input: &'a str) -> PResult<'a, String> {
-    let (mut rest, first) = parse_ident_with_hyphens(input)?;
-    let mut full = first.to_string();
-    while let Some(after) = rest.strip_prefix("::") {
-        if after.starts_with('(') {
-            break;
-        }
-        if let Ok((r2, part)) = parse_ident_with_hyphens(after) {
-            full.push_str("::");
-            full.push_str(part);
-            rest = r2;
-        } else {
-            break;
-        }
-    }
-    Ok((rest, full))
-}
-
-fn parse_symbolic_deref_segments(mut rest: &str, mut combined: Expr) -> PResult<'_, Expr> {
-    loop {
-        if let Some(after_sep) = rest.strip_prefix("::(") {
-            combined = concat_exprs(combined, Expr::Literal(Value::str("::".to_string())));
-            let (after_expr, seg_expr) = expression(after_sep)?;
-            let (after_expr, _) = ws(after_expr)?;
-            let (r, _) = parse_char(after_expr, ')')?;
-            combined = concat_exprs(combined, seg_expr);
-            rest = r;
-        } else if let Some(after_sep) = rest.strip_prefix("::") {
-            if let Ok((r, ident)) = parse_ident_with_hyphens(after_sep) {
-                let sep_ident = format!("::{}", ident);
-                combined = concat_exprs(combined, Expr::Literal(Value::str(sep_ident)));
-                rest = r;
-            } else {
-                break;
-            }
-        } else {
-            break;
-        }
-    }
-    Ok((rest, combined))
-}
-
-/// Parse a variable name from raw string (used in interpolation).
-pub(super) fn parse_var_name_from_str(input: &str) -> (&str, String) {
-    // Handle twigils: $*, $?, $!, $^
-    let (rest, twigil) = if input.starts_with('*')
-        || input.starts_with('?')
-        || input.starts_with('!')
-        || input.starts_with('^')
-    {
-        (&input[1..], &input[..1])
-    } else {
-        (input, "")
-    };
-    let (rest_after_name, name) = parse_interpolation_qualified_ident_with_hyphens_or_empty(rest);
-    let (rest_after_name, name) = parse_var_name_adverb_suffixes(rest_after_name, name);
-    let full_name = if twigil.is_empty() {
-        name
-    } else {
-        format!("{}{}", twigil, name)
-    };
-    (rest_after_name, full_name)
-}
-
-/// Parse a $variable reference.
-pub(super) fn scalar_var(input: &str) -> PResult<'_, Expr> {
-    let (input, _) = parse_char(input, '$')?;
-    // Detect Perl 5 special variables and throw X::Syntax::Perl5Var errors.
-    if let Some(err) = detect_perl5_scalar_var(input) {
-        return Err(PError::fatal(err));
-    }
-    // Handle $(stmt; expr) — statement block in scalar context
-    // Try to parse as a statement list first (for cases like `$(let $a = 23; $a)`)
-    if let Some(inner) = input.strip_prefix('(') {
-        // Handle $(;) — semicolons only inside $() produce an empty itemized list
-        {
-            let mut r = inner;
-            loop {
-                r = r.trim_start();
-                if let Some(rest) = r.strip_prefix(';') {
-                    r = rest;
-                } else {
-                    break;
-                }
-            }
-            r = r.trim_start();
-            if let Some(rest) = r.strip_prefix(')') {
-                return Ok((
-                    rest,
-                    Expr::MethodCall {
-                        target: Box::new(Expr::ArrayLiteral(Vec::new())),
-                        name: Symbol::intern("item"),
-                        args: vec![],
-                        modifier: None,
-                        quoted: false,
-                    },
-                ));
-            }
-        }
-        if let Ok(result) = parse_dollar_paren_block(inner) {
-            return Ok(result);
-        }
-        return paren_expr(input);
-    }
-    // Handle $@array (array in item context) and $%hash (hash in item context)
-    if input.starts_with('@') || input.starts_with('%') {
-        let sigil = &input[..1];
-        let after_sigil = &input[1..];
-        // Check for twigils or plain identifier
-        let (rest, twigil) = if after_sigil.starts_with('*')
-            || after_sigil.starts_with('?')
-            || after_sigil.starts_with('!')
-        {
-            (&after_sigil[1..], &after_sigil[..1])
-        } else {
-            (after_sigil, "")
-        };
-        if let Ok((rest, name)) = parse_qualified_ident_with_hyphens(rest) {
-            let full_name = if twigil.is_empty() {
-                name
-            } else {
-                format!("{}{}", twigil, name)
-            };
-            let inner = if sigil == "@" {
-                Expr::ArrayVar(full_name)
-            } else {
-                Expr::HashVar(full_name)
-            };
-            return Ok((rest, Expr::Itemize(Box::new(inner))));
-        }
-    }
-    // Handle nested scalar dereference syntax ($$x / $&f) by parsing the
-    // inner variable term. This keeps '$' from being misparsed as an
-    // anonymous bare-sigil state variable in expressions like `is($$o, ...)`.
-    if input.starts_with('$') {
-        return scalar_var(input);
-    }
-    if input.starts_with('&') {
-        return code_var(input);
-    }
-    // Handle $_ special variable
-    if input.starts_with('_')
-        && (input.len() == 1
-            || !input[1..]
-                .chars()
-                .next()
-                .is_some_and(is_raku_identifier_continue))
-    {
-        return Ok((&input[1..], Expr::Var("_".to_string())));
-    }
-    // Handle $/ (match variable)
-    if let Some(stripped) = input.strip_prefix('/') {
-        return Ok((stripped, Expr::Var("/".to_string())));
-    }
-    // Handle $¢ (current match cursor variable)
-    if let Some(stripped) = input.strip_prefix('\u{00A2}') {
-        return Ok((stripped, Expr::Var("\u{00A2}".to_string())));
-    }
-    // Handle $! (exception variable) — bare $! without name after it
-    if let Some(after) = input.strip_prefix('!') {
-        // If next char is alphanumeric or _, it's a twigil (e.g. $!attr)
-        // If not, it's the bare $! variable
-        let is_twigil =
-            !after.is_empty() && after.chars().next().is_some_and(is_raku_identifier_start);
-        if !is_twigil {
-            return Ok((after, Expr::Var("!".to_string())));
-        }
-    }
-    // Handle $<name> (named capture variable), including forms like $<&foo>.
-    if let Some(after_lt) = input.strip_prefix('<')
-        && let Some(end) = after_lt.find('>')
-    {
-        let name = &after_lt[..end];
-        if !name.is_empty() {
-            let after_gt = &after_lt[end + 1..];
-            return Ok((after_gt, Expr::CaptureVar(name.to_string())));
-        }
-        // `$<>` is the zen slice of the match variable: sugar for `$/<>`.
-        // Return `$/` and leave the `<>` for postfix zen-angle parsing.
-        return Ok((input, Expr::Var("/".to_string())));
-    }
-    // Handle $=finish and other Pod variables ($=pod, $=data, etc.)
-    if let Some(after_eq) = input.strip_prefix('=') {
-        let (rest, name) = parse_ident_with_hyphens(after_eq)?;
-        let full_name = format!("={}", name);
-        return Ok((rest, Expr::Var(full_name)));
-    }
-    // Symbolic variable dereference: $::("name"), $::("pkg")::name::($dyn)
-    if let Some(after_colons) = input.strip_prefix("::(") {
-        let (after_expr, first_expr) = expression(after_colons)?;
-        let (after_expr, _) = ws(after_expr)?;
-        let (rest, _) = parse_char(after_expr, ')')?;
-        let (rest, combined) = parse_symbolic_deref_segments(rest, first_expr)?;
-        return Ok((
-            rest,
-            Expr::SymbolicDeref {
-                sigil: "$".to_string(),
-                expr: Box::new(combined),
-            },
-        ));
-    }
-    if let Some(after_brace) = input.strip_prefix("::{") {
-        let (rest, _) = ws(after_brace)?;
-        let (rest, key_expr) = expression(rest)?;
-        let (rest, _) = ws(rest)?;
-        let (rest, _) = parse_char(rest, '}')?;
-        return Ok((
-            rest,
-            Expr::Index {
-                target: Box::new(Expr::PseudoStash("::".to_string())),
-                index: Box::new(key_expr),
-                is_positional: false,
-            },
-        ));
-    }
-    if let Some(after_bracket) = input.strip_prefix("::<")
-        && let Some(end) = after_bracket.find('>')
-    {
-        let symbol = &after_bracket[..end];
-        if !symbol.is_empty() {
-            return Ok((
-                &after_bracket[end + 1..],
-                Expr::Index {
-                    target: Box::new(Expr::PseudoStash("::".to_string())),
-                    index: Box::new(Expr::Literal(Value::str(symbol.to_string()))),
-                    is_positional: false,
-                },
-            ));
-        }
-    }
-    // Named parameter variable inside blocks: $:name
-    if let Some(after_colon) = input.strip_prefix(':') {
-        let (rest, name) = parse_ident_with_hyphens(after_colon)?;
-        return Ok((rest, Expr::Var(format!(":{}", name))));
-    }
-    // Handle bare $ (anonymous state variable)
-    // If next char is not a valid identifier start, twigil, or special char, it's an anonymous var
-    let next_is_ident_or_twigil = !input.is_empty() && {
-        let first_char = input.chars().next().unwrap();
-        is_raku_identifier_start(first_char)
-            || first_char.is_numeric()
-            || first_char == '*'
-            || first_char == '?'
-            || first_char == '!'
-            || first_char == '^'
-            || first_char == '.'
-            || first_char == '~'
-    };
-    if !next_is_ident_or_twigil {
-        let id = ANON_STATE_COUNTER.fetch_add(1, Ordering::Relaxed);
-        return Ok((input, Expr::Var(format!("__ANON_STATE_{id}__"))));
-    }
-    // $. followed by non-alphabetic: shorthand for self. (e.g., $.^name = self.^name)
-    // Return "self" as a BareWord and leave the '.' for postfix parsing.
-    if input.starts_with('.')
-        && input.len() > 1
-        && !input.as_bytes()[1].is_ascii_alphabetic()
-        && input.as_bytes()[1] != b'_'
-    {
-        return Ok((input, Expr::BareWord("self".to_string())));
-    }
-    // $.ident followed by ( or : → parse as self.ident(args) or self.ident: args
-    // This ensures $.method(args) works as a method call, not a variable call.
-    if input.starts_with('.')
-        && input.len() > 1
-        && (input.as_bytes()[1].is_ascii_alphabetic() || input.as_bytes()[1] == b'_')
-    {
-        // Look ahead past the identifier to check for ( or :
-        let after_dot = &input[1..];
-        let ident_end = after_dot
-            .find(|c: char| !is_raku_identifier_continue(c) && c != '-')
-            .unwrap_or(after_dot.len());
-        let after_ident = &after_dot[ident_end..];
-        if after_ident.starts_with('(')
-            || (after_ident.starts_with(':') && !after_ident.starts_with("::"))
-        {
-            return Ok((input, Expr::BareWord("self".to_string())));
-        }
-    }
-    // Handle twigils: $*FOO, $?FILE, $!attr, $.attr, $~MAIN
-    let (rest, twigil) = if input.starts_with('*')
-        || input.starts_with('?')
-        || input.starts_with('!')
-        || input.starts_with('^')
-        || (input.starts_with('.')
-            && input.len() > 1
-            && (input.as_bytes()[1].is_ascii_alphabetic() || input.as_bytes()[1] == b'_'))
-        || (input.starts_with('~') && input.len() > 1 && input.as_bytes()[1].is_ascii_alphabetic())
-    {
-        (&input[1..], &input[..1])
-    } else {
-        (input, "")
-    };
-    let (rest, name) = parse_qualified_ident_prefix_with_hyphens(rest)?;
-    if rest.starts_with("::(") {
-        let (rest, combined) =
-            parse_symbolic_deref_segments(rest, Expr::Literal(Value::str(name.clone())))?;
-        return Ok((
-            rest,
-            Expr::SymbolicDeref {
-                sigil: "$".to_string(),
-                expr: Box::new(combined),
-            },
-        ));
-    }
-    let (rest, name) = parse_var_name_adverb_suffixes(rest, name);
-    let full_name = if twigil.is_empty() {
-        name
-    } else {
-        format!("{}{}", twigil, name)
-    };
-    // $?LINE is a compile-time constant: replace with the current line number
-    if full_name == "?LINE" {
-        let line = current_line_number(input);
-        return Ok((rest, Expr::Literal(Value::Int(line))));
-    }
-    // $?TABSTOP is a compile-time constant: hardcoded to 8
-    if full_name == "?TABSTOP" {
-        return Ok((rest, Expr::Literal(Value::Int(8))));
-    }
-    // Normalize positional capture variables: $00 → $0, $01 → $1, etc.
-    // Strip leading zeros from all-digit variable names so $00 resolves to $0.
-    let full_name = if !full_name.is_empty()
-        && full_name.chars().next().unwrap().is_ascii_digit()
-        && full_name.chars().all(|c| c.is_ascii_digit())
-    {
-        let trimmed = full_name.trim_start_matches('0');
-        if trimmed.is_empty() {
-            "0".to_string()
-        } else {
-            trimmed.to_string()
-        }
-    } else {
-        full_name
-    };
-    Ok((rest, Expr::Var(full_name)))
-}
-
-/// Parse identifier allowing kebab-case hyphens and apostrophes (e.g., `my-var`, `same'proto`).
-/// A hyphen/apostrophe is only part of the name if followed by an alphabetic char or `_`,
-/// so `$pd--` parses as `$pd` + postfix `--`.
-pub(crate) fn parse_ident_with_hyphens<'a>(input: &'a str) -> PResult<'a, &'a str> {
-    let first = input
-        .chars()
-        .next()
-        .ok_or_else(|| PError::expected("identifier"))?;
-    if !is_raku_identifier_start(first) && !first.is_numeric() {
-        return Err(PError::expected("identifier"));
-    }
-    let mut end = first.len_utf8();
-    for c in input[end..].chars() {
-        if is_raku_identifier_continue(c) {
-            end += c.len_utf8();
-        } else {
-            break;
-        }
-    }
-    loop {
-        let remaining = &input[end..];
-        let sep = if remaining.starts_with('-') {
-            Some('-')
-        } else if remaining.starts_with('\'') {
-            Some('\'')
-        } else {
-            None
-        };
-        if let Some(sep) = sep {
-            let after_sep = &remaining[sep.len_utf8()..];
-            if let Some(next) = after_sep.chars().next()
-                && is_raku_identifier_start(next)
-            {
-                end += sep.len_utf8();
-                for c in after_sep.chars() {
-                    if is_raku_identifier_continue(c) {
-                        end += c.len_utf8();
-                    } else {
-                        break;
-                    }
-                }
-                continue;
-            }
-        }
-        break;
-    }
-    Ok((&input[end..], &input[..end]))
-}
-
-fn parse_qualified_ident_with_hyphens<'a>(input: &'a str) -> PResult<'a, String> {
-    let (mut rest, first) = parse_ident_with_hyphens(input)?;
-    let mut full = first.to_string();
-    while let Some(after) = rest.strip_prefix("::") {
-        let (r2, part) = parse_ident_with_hyphens(after)?;
-        full.push_str("::");
-        full.push_str(part);
-        rest = r2;
-    }
-    Ok((rest, normalize_raku_identifier(&full)))
-}
-
-fn parse_var_name_adverb_suffixes(mut rest: &str, mut name: String) -> (&str, String) {
-    while rest.starts_with(':') && !rest.starts_with("::") {
-        let after_colon = &rest[1..];
-        if let Ok((r2, suffix)) = parse_ident_with_hyphens(after_colon) {
-            // Keep postfix adverb names available to postfix parsing.
-            // This avoids treating `$a:delete` as a variable named `a:delete`.
-            if matches!(suffix, "delete" | "exists" | "v" | "kv" | "k" | "p") {
-                break;
-            }
-            name.push(':');
-            name.push_str(suffix);
-            rest = r2;
-            // Parse adverb value: <...>, «...», [...], (...)
-            if let Some(canonical) = parse_adverb_value(rest) {
-                name.push_str(&canonical.0);
-                rest = canonical.1;
-            }
-        } else {
-            break;
-        }
-    }
-    (rest, name)
-}
-
-/// Public wrapper for `parse_adverb_value` used by `stmt/mod.rs`.
-pub(crate) fn parse_adverb_value_pub(input: &str) -> Option<(String, &str)> {
-    parse_adverb_value(input)
-}
-
-/// Parse adverb value brackets and canonicalize to `<word1 word2>` form.
-/// Returns (canonical_string, remaining_input) or None if no adverb value follows.
-fn parse_adverb_value(input: &str) -> Option<(String, &str)> {
-    let first_char = input.chars().next()?;
-    match first_char {
-        '<' => {
-            // Angle brackets: <a b>
-            let close = input.find('>')?;
-            let content = &input[1..close];
-            let rest = &input[close + 1..];
-            Some((format!("<{}>", content), rest))
-        }
-        '\u{00AB}' => {
-            // French quotes: « »
-            let close_char = '\u{00BB}';
-            let close = input[first_char.len_utf8()..].find(close_char)?;
-            let content = &input[first_char.len_utf8()..first_char.len_utf8() + close];
-            let rest = &input[first_char.len_utf8() + close + close_char.len_utf8()..];
-            Some((format!("<{}>", content), rest))
-        }
-        '[' => {
-            // Square brackets: ['a','b']
-            let close = find_matching_bracket(input, '[', ']')?;
-            let content = &input[1..close];
-            let rest = &input[close + 1..];
-            let words = parse_comma_separated_values(content);
-            Some((format!("<{}>", words.join(" ")), rest))
-        }
-        '(' => {
-            // Parentheses: ('a','b')
-            let close = find_matching_bracket(input, '(', ')')?;
-            let content = &input[1..close];
-            let rest = &input[close + 1..];
-            let words = parse_comma_separated_values(content);
-            Some((format!("<{}>", words.join(" ")), rest))
-        }
-        _ => None,
-    }
-}
-
-/// Find the position of a matching closing bracket, handling nesting.
-fn find_matching_bracket(input: &str, open: char, close: char) -> Option<usize> {
-    let mut depth = 0;
-    let mut in_string = false;
-    let mut string_char = ' ';
-    for (i, c) in input.char_indices() {
-        if in_string {
-            if c == string_char {
-                in_string = false;
-            }
-            continue;
-        }
-        if c == '\'' || c == '"' {
-            in_string = true;
-            string_char = c;
-        } else if c == open {
-            depth += 1;
-        } else if c == close {
-            depth -= 1;
-            if depth == 0 {
-                return Some(i);
-            }
-        }
-    }
-    None
-}
-
-/// Parse comma-separated values, stripping quotes.
-/// `'a','b'` → vec!["a", "b"]
-fn parse_comma_separated_values(input: &str) -> Vec<String> {
-    input
-        .split(',')
-        .map(|s| {
-            let s = s.trim();
-            // Strip surrounding quotes
-            if (s.starts_with('\'') && s.ends_with('\''))
-                || (s.starts_with('"') && s.ends_with('"'))
-            {
-                s[1..s.len() - 1].to_string()
-            } else {
-                s.to_string()
-            }
-        })
-        .collect()
-}
-
-fn parse_interpolation_qualified_ident_with_hyphens_or_empty(input: &str) -> (&str, String) {
-    // Interpolation variable names should stop before Unicode superscript
-    // exponent markers (e.g. "$no²" -> var "no", then literal "²").
-    let mut rest = input;
-    let mut full = String::new();
-    let mut parsed_any = false;
-    while let Some(first) = rest.chars().next() {
-        if !(first.is_ascii_alphabetic() || first == '_') {
-            break;
-        }
-        let mut end = first.len_utf8();
-        let rest_bytes = rest.as_bytes();
-        for c in rest[end..].chars() {
-            if c.is_ascii_alphanumeric() || c == '_' {
-                end += c.len_utf8();
-            } else if c == '-' {
-                // Kebab-case: hyphen is part of the identifier only when
-                // followed by an alphabetic character (e.g. `$my-var`).
-                // A trailing hyphen (`$a-`) is NOT part of the name.
-                let next_pos = end + 1; // '-' is ASCII, always 1 byte
-                if next_pos < rest_bytes.len() && rest_bytes[next_pos].is_ascii_alphabetic() {
-                    end += c.len_utf8();
-                } else {
-                    break;
-                }
-            } else {
-                break;
-            }
-        }
-        if parsed_any {
-            full.push_str("::");
-        }
-        full.push_str(&rest[..end]);
-        rest = &rest[end..];
-        parsed_any = true;
-        if let Some(after_scope) = rest.strip_prefix("::") {
-            rest = after_scope;
-        } else {
-            break;
-        }
-    }
-    if parsed_any {
-        (rest, full)
-    } else {
-        (input, String::new())
-    }
-}
-/// Parse an @array variable reference.
-pub(super) fn array_var(input: &str) -> PResult<'_, Expr> {
-    let (input, _) = parse_char(input, '@')?;
-    // Detect Perl 5 special array variables (@-, @+) and throw X::Syntax::Perl5Var.
-    if let Some(err) = detect_perl5_sigil_var('@', input) {
-        return Err(PError::fatal(err));
-    }
-    // Symbolic variable dereference: @::("name")
-    if let Some(after_colons) = input.strip_prefix("::(") {
-        let (after_expr, expr) = expression(after_colons)?;
-        let (after_expr, _) = ws(after_expr)?;
-        let (rest, _) = parse_char(after_expr, ')')?;
-        return Ok((
-            rest,
-            Expr::SymbolicDeref {
-                sigil: "@".to_string(),
-                expr: Box::new(expr),
-            },
-        ));
-    }
-    // Handle `.` twigil: @.attr → self.attr (in list context)
-    if input.starts_with('.')
-        && input.len() > 1
-        && (input.as_bytes()[1].is_ascii_alphabetic() || input.as_bytes()[1] == b'_')
-    {
-        let after_dot = &input[1..];
-        let (rest, name) = parse_qualified_ident_with_hyphens(after_dot)?;
-        return Ok((rest, Expr::ArrayVar(format!(".{}", name))));
-    }
-    // Handle twigils
-    let (rest, twigil) = if input.starts_with('*')
-        || input.starts_with('?')
-        || input.starts_with('!')
-        || input.starts_with('^')
-        || (input.starts_with('~') && input.len() > 1 && input.as_bytes()[1].is_ascii_alphabetic())
-    {
-        (&input[1..], &input[..1])
-    } else {
-        (input, "")
-    };
-    // Contextualized scalar specials (e.g., @$/, @$_): parse `$...` then lift
-    // to an array variable targeting the same underlying name.
-    if twigil.is_empty() && rest.starts_with('$') {
-        // Handle @$<name> — list coercion of a named capture variable
-        if let Ok((r2, expr @ Expr::CaptureVar(_))) = scalar_var(rest) {
-            return Ok((
-                r2,
-                Expr::MethodCall {
-                    target: Box::new(expr),
-                    name: crate::symbol::Symbol::intern("list"),
-                    args: vec![],
-                    modifier: None,
-                    quoted: false,
-                },
-            ));
-        }
-        if let Ok((r2, Expr::Var(name))) = scalar_var(rest) {
-            return Ok((r2, Expr::ArrayVar(name)));
-        }
-    }
-    // @{expr} is Perl 5 array dereference syntax — throw X::Obsolete
-    if twigil.is_empty()
-        && rest.starts_with('{')
-        && let Ok((_r2, inner)) = super::misc::block_or_hash_expr(rest)
-        && !matches!(inner, crate::ast::Expr::Hash(_))
-    {
-        return Err(PError::fatal(
-            "X::Obsolete: Unsupported use of @{expr}. In Raku please use: @(expr).".to_string(),
-        ));
-    }
-    // Handle @<name> — list coercion of a named capture variable from $/
-    // (e.g., after a regex match, @<fie> gives the positional elements of $<fie>)
-    if twigil.is_empty()
-        && rest.starts_with('<')
-        && let Some(after_lt) = rest.strip_prefix('<')
-        && let Some(end) = after_lt.find('>')
-    {
-        let name = &after_lt[..end];
-        if !name.is_empty() {
-            let after_gt = &after_lt[end + 1..];
-            return Ok((
-                after_gt,
-                Expr::MethodCall {
-                    target: Box::new(Expr::CaptureVar(name.to_string())),
-                    name: crate::symbol::Symbol::intern("list"),
-                    args: vec![],
-                    modifier: None,
-                    quoted: false,
-                },
-            ));
-        }
-    }
-    // Bare @ (anonymous array variable) — each occurrence gets a unique name
-    let next_is_ident =
-        !rest.is_empty() && rest.chars().next().is_some_and(is_raku_identifier_start);
-    if !next_is_ident && twigil.is_empty() {
-        let id = ANON_ARRAY_COUNTER.fetch_add(1, Ordering::Relaxed);
-        return Ok((rest, Expr::ArrayVar(format!("__ANON_ARRAY_{id}__"))));
-    }
-    let (rest, name) = parse_qualified_ident_with_hyphens(rest)?;
-    let (rest, name) = parse_var_name_adverb_suffixes(rest, name);
-    let full_name = if twigil.is_empty() {
-        name
-    } else {
-        format!("{}{}", twigil, name)
-    };
-    Ok((rest, Expr::ArrayVar(full_name)))
-}
-
-/// Parse a %hash variable reference.
-pub(super) fn hash_var(input: &str) -> PResult<'_, Expr> {
-    let (input, _) = parse_char(input, '%')?;
-    // Detect Perl 5 special hash variables (%-, %+, %!) and throw X::Syntax::Perl5Var.
-    if let Some(err) = detect_perl5_sigil_var('%', input) {
-        return Err(PError::fatal(err));
-    }
-    // Symbolic variable dereference: %::("name")
-    if let Some(after_colons) = input.strip_prefix("::(") {
-        let (after_expr, expr) = expression(after_colons)?;
-        let (after_expr, _) = ws(after_expr)?;
-        let (rest, _) = parse_char(after_expr, ')')?;
-        return Ok((
-            rest,
-            Expr::SymbolicDeref {
-                sigil: "%".to_string(),
-                expr: Box::new(expr),
-            },
-        ));
-    }
-    // Handle `.` twigil: %.attr → self.attr (in hash context)
-    if input.starts_with('.')
-        && input.len() > 1
-        && (input.as_bytes()[1].is_ascii_alphabetic() || input.as_bytes()[1] == b'_')
-    {
-        let after_dot = &input[1..];
-        let (rest, name) = parse_qualified_ident_with_hyphens(after_dot)?;
-        return Ok((rest, Expr::HashVar(format!(".{}", name))));
-    }
-    // Handle twigils
-    let (rest, twigil) = if input.starts_with('*')
-        || input.starts_with('?')
-        || input.starts_with('!')
-        || input.starts_with('^')
-        || (input.starts_with('~') && input.len() > 1 && input.as_bytes()[1].is_ascii_alphabetic())
-    {
-        (&input[1..], &input[..1])
-    } else {
-        (input, "")
-    };
-    // Contextualized scalar specials (e.g., %$h, %$/, %$_): parse `$...` then
-    // coerce to hash context via a `.hash` method call.
-    if twigil.is_empty()
-        && rest.starts_with('$')
-        && let Ok((r2, expr)) = scalar_var(rest)
-    {
-        return Ok((
-            r2,
-            Expr::MethodCall {
-                target: Box::new(expr),
-                name: crate::symbol::Symbol::intern("hash"),
-                args: vec![],
-                modifier: None,
-                quoted: false,
-            },
-        ));
-    }
-    // %{...} — hash coercion of a block expression
-    // In Raku, %{expr} treats the block as a hash initializer.
-    if twigil.is_empty()
-        && rest.starts_with('{')
-        && let Ok((r2, inner)) = super::misc::block_or_hash_expr(rest)
-    {
-        // Wrap in hash() call — this will fail at runtime if odd number of elements
-        return Ok((
-            r2,
-            Expr::Call {
-                name: crate::symbol::Symbol::intern("hash"),
-                args: vec![inner],
-            },
-        ));
-    }
-    // Handle %<name> — hash coercion of a named capture variable from $/
-    // (e.g., after a regex match, %<foe> gives the hash view of $<foe>)
-    if twigil.is_empty()
-        && rest.starts_with('<')
-        && let Some(after_lt) = rest.strip_prefix('<')
-        && let Some(end) = after_lt.find('>')
-    {
-        let name = &after_lt[..end];
-        if !name.is_empty() {
-            let after_gt = &after_lt[end + 1..];
-            return Ok((
-                after_gt,
-                Expr::MethodCall {
-                    target: Box::new(Expr::CaptureVar(name.to_string())),
-                    name: crate::symbol::Symbol::intern("hash"),
-                    args: vec![],
-                    modifier: None,
-                    quoted: false,
-                },
-            ));
-        }
-    }
-    // Bare % (anonymous hash variable)
-    let next_is_ident =
-        !rest.is_empty() && rest.chars().next().is_some_and(is_raku_identifier_start);
-    if !next_is_ident && twigil.is_empty() {
-        return Ok((rest, Expr::HashVar("__ANON_HASH__".to_string())));
-    }
-    // Special: %*ENV
-    let (rest, name) = parse_qualified_ident_with_hyphens(rest)?;
-    let (rest, name) = parse_var_name_adverb_suffixes(rest, name);
-    let full_name = if twigil.is_empty() {
-        name
-    } else {
-        format!("{}{}", twigil, name)
-    };
-    Ok((rest, Expr::HashVar(full_name)))
-}
-
-/// Check if an identifier is a known pseudo-package name.
-pub(crate) fn is_pseudo_package(name: &str) -> bool {
-    matches!(
-        name,
-        "SETTING"
-            | "CALLER"
-            | "CALLERS"
-            | "OUTER"
-            | "OUTERS"
-            | "CORE"
-            | "GLOBAL"
-            | "MY"
-            | "OUR"
-            | "DYNAMIC"
-            | "UNIT"
-            | "LEXICAL"
-            | "CLIENT"
-            | "PROCESS"
-            | "COMPILING"
-    )
-}
-
-/// Parse a &code variable reference.
-/// Handles `&foo`, twigil forms `&?BLOCK`, operator references like `&infix:<+>`,
-/// package-qualified code refs like `&SETTING::OUTER::name`, and indirect package
-/// lookups like `&::($expr)::name` or `&CALLER::($expr)::name`.
-pub(super) fn code_var(input: &str) -> PResult<'_, Expr> {
-    let (input, _) = parse_char(input, '&')?;
-    // Dereference callable stored in a variable/expression:
-    // &$x, &@x, &%x, &($expr)
-    if input.starts_with('$') {
-        return scalar_var(input);
-    }
-    if input.starts_with('@') {
-        return array_var(input);
-    }
-    if input.starts_with('%') {
-        return hash_var(input);
-    }
-    if let Some(after_paren) = input.strip_prefix('(') {
-        let (rest, expr) = expression(after_paren)?;
-        let (rest, _) = ws(rest)?;
-        let (rest, _) = parse_char(rest, ')')?;
-        return Ok((rest, expr));
-    }
-    // Callable block literal dereference: &{ ... }
-    if input.starts_with('{') {
-        let (rest, body) = parse_block_body(input)?;
-        return Ok((
-            rest,
-            Expr::AnonSub {
-                body,
-                is_rw: false,
-                is_block: true,
-            },
-        ));
-    }
-    // Handle &[op] — short form for &infix:<op>
-    if let Some(after_bracket) = input.strip_prefix('[')
-        && let Some(end_pos) = after_bracket.find(']')
-    {
-        let op_name = &after_bracket[..end_pos];
-        let rest = &after_bracket[end_pos + 1..];
-        let full_name = format!("infix:<{}>", op_name);
-        return Ok((rest, Expr::CodeVar(full_name)));
-    }
-    // Handle &CALLER::($expr)::name — CALLER prefix followed by indirect lookup
-    if let Some(after_caller) = input.strip_prefix("CALLER::(") {
-        let (after_expr, expr) = expression(after_caller)?;
-        let (after_expr, _) = parse_char(after_expr, ')')?;
-        let (after_expr, _) = ws(after_expr)?;
-        if let Some(after_sep) = after_expr.strip_prefix("::") {
-            let (rest, name) = parse_ident_with_hyphens(after_sep)?;
-            return Ok((
-                rest,
-                Expr::IndirectCodeLookup {
-                    package: Box::new(expr),
-                    name: normalize_raku_identifier(name),
-                },
-            ));
-        }
-        return Err(PError::expected(":: after indirect package lookup"));
-    }
-    // Handle &::($expr)::name — indirect package lookup
-    // Handle &::($expr) — symbolic code dereference (looks up &name at runtime)
-    if let Some(after_colons) = input.strip_prefix("::(") {
-        let (after_expr, expr) = expression(after_colons)?;
-        let (after_expr, _) = parse_char(after_expr, ')')?;
-        let (after_expr, _) = ws(after_expr)?;
-        if let Some(after_sep) = after_expr.strip_prefix("::") {
-            let (rest, name) = parse_ident_with_hyphens(after_sep)?;
-            return Ok((
-                rest,
-                Expr::IndirectCodeLookup {
-                    package: Box::new(expr),
-                    name: normalize_raku_identifier(name),
-                },
-            ));
-        }
-        // No ::name suffix — this is &::("name"), a symbolic code variable deref
-        return Ok((
-            after_expr,
-            Expr::SymbolicDeref {
-                sigil: "&".to_string(),
-                expr: Box::new(expr),
-            },
-        ));
-    }
-    // Handle package-qualified code refs: &SETTING::OUTER::...::name
-    // or &CALLER::SETTING::OUTER::...::name
-    if let Ok((_, first_ident)) = parse_ident_with_hyphens(input)
-        && is_pseudo_package(first_ident)
-    {
-        let after_first = &input[first_ident.len()..];
-        if after_first.starts_with("::") {
-            // Consume the full Package::Package::...::name chain
-            let mut pos = 0;
-            let mut qualified = String::new();
-            loop {
-                if let Ok((_, ident)) = parse_ident_with_hyphens(&input[pos..]) {
-                    let after_ident = &input[pos + ident.len()..];
-                    if after_ident.starts_with("::") && is_pseudo_package(ident) {
-                        // This is a pseudo-package, consume it and the ::
-                        qualified.push_str(ident);
-                        qualified.push_str("::");
-                        pos += ident.len() + 2;
-                        continue;
-                    }
-                    // This is the final name
-                    qualified.push_str(&normalize_raku_identifier(ident));
-                    pos += ident.len();
-                    return Ok((&input[pos..], Expr::CodeVar(qualified)));
-                }
-                break;
-            }
-        }
-    }
-    // Handle package-qualified code refs: &ClassName::method, &Foo::Bar::method, etc.
-    if let Ok((_, first_ident)) = parse_ident_with_hyphens(input) {
-        let after_first = &input[first_ident.len()..];
-        if after_first.starts_with("::") {
-            let mut pos = 0;
-            let mut parts: Vec<String> = Vec::new();
-            // Consume all Ident:: segments
-            loop {
-                if let Ok((_, ident)) = parse_ident_with_hyphens(&input[pos..]) {
-                    let after_ident = &input[pos + ident.len()..];
-                    if after_ident.starts_with("::") {
-                        parts.push(normalize_raku_identifier(ident));
-                        pos += ident.len() + 2; // skip ident and ::
-                        continue;
-                    }
-                    // Final identifier (no :: after it)
-                    parts.push(normalize_raku_identifier(ident));
-                    pos += ident.len();
-                    if parts.len() >= 2 {
-                        let qualified = parts.join("::");
-                        return Ok((&input[pos..], Expr::CodeVar(qualified)));
-                    }
-                }
-                break;
-            }
-        }
-    }
-    // Handle twigils on code vars: &?BLOCK, &!DISPATCHER, &^x, &*FOO
-    let (rest, twigil) = if let Some(stripped) = input.strip_prefix('?') {
-        (stripped, "?")
-    } else if let Some(stripped) = input.strip_prefix('!') {
-        (stripped, "!")
-    } else if let Some(stripped) = input.strip_prefix('^') {
-        (stripped, "^")
-    } else if let Some(stripped) = input.strip_prefix('*') {
-        (stripped, "*")
-    } else {
-        (input, "")
-    };
-    if twigil.is_empty()
-        && let Ok((op_rest, op_name)) = super::super::stmt::parse_sub_name_pub(rest)
-        && matches!(
-            op_name.as_str(),
-            n if n.starts_with("infix:<")
-                || n.starts_with("prefix:<")
-                || n.starts_with("postfix:<")
-                || n.starts_with("term:<")
-                || n.starts_with("circumfix:<")
-                || n.starts_with("postcircumfix:<")
-        )
-    {
-        return Ok((op_rest, Expr::CodeVar(op_name)));
-    }
-
-    let (rest, name) = parse_ident_with_hyphens(rest)?;
-    // Check for operator reference:
-    // &infix:<OP>, &infix:<<OP>>, &infix:«OP», &[op], ...
-    if twigil.is_empty()
-        && matches!(
-            name,
-            "infix" | "prefix" | "postfix" | "term" | "circumfix" | "postcircumfix"
-        )
-        && let Some((r, full_name)) = parse_operator_code_ref_suffix(name, rest)
-    {
-        return Ok((r, Expr::CodeVar(full_name)));
-    }
-    let full_name = if twigil.is_empty() {
-        normalize_raku_identifier(name)
-    } else {
-        format!("{}{}", twigil, normalize_raku_identifier(name))
-    };
-    Ok((rest, Expr::CodeVar(full_name)))
-}
-
-fn parse_operator_code_ref_suffix<'a>(category: &str, rest: &'a str) -> Option<(&'a str, String)> {
-    if let Some(after_open) = rest.strip_prefix(":<<")
-        && let Some(end_pos) = after_open.find(">>")
-    {
-        let symbol = after_open[..end_pos].trim();
-        let after_close = &after_open[end_pos + 2..];
-        return Some((after_close, format!("{category}:<{symbol}>")));
-    }
-
-    if let Some(after_open) = rest.strip_prefix(":<") {
-        // Nested <...> is valid in categorical operator names.
-        let mut depth = 1u32;
-        let mut chars = after_open.char_indices();
-        while let Some((i, c)) = chars.next() {
-            match c {
-                '>' => {
-                    depth -= 1;
-                    if depth == 0 {
-                        let symbol = &after_open[..i];
-                        let after_close = &after_open[i + 1..];
-                        return Some((after_close, format!("{category}:<{symbol}>")));
-                    }
-                }
-                '<' => depth += 1,
-                '\\' => {
-                    chars.next();
-                }
-                _ => {}
-            }
-        }
-    }
-
-    if let Some(after_open) = rest.strip_prefix(":\u{ab}")
-        && let Some(end_pos) = after_open.find('\u{bb}')
-    {
-        let symbol = after_open[..end_pos].trim();
-        let after_close = &after_open[end_pos + '\u{bb}'.len_utf8()..];
-        return Some((after_close, format!("{category}:<{symbol}>")));
-    }
-
-    if let Some(after_open) = rest.strip_prefix(":['")
-        && let Some(end_pos) = after_open.find("']")
-    {
-        let symbol = after_open[..end_pos]
-            .replace("\\'", "'")
-            .replace("\\\\", "\\");
-        let after_close = &after_open[end_pos + 2..];
-        return Some((after_close, format!("{category}:<{symbol}>")));
-    }
-
-    if let Some(after_open) = rest.strip_prefix(":[\"")
-        && let Some(end_pos) = after_open.find("\"]")
-    {
-        let symbol = &after_open[..end_pos];
-        let after_close = &after_open[end_pos + 2..];
-        return Some((after_close, format!("{category}:<{symbol}>")));
-    }
-
-    None
-}
-
-/// Parse `$(stmt; stmt; ... expr)` — statement block in scalar context.
-/// Returns DoBlock expression.
-pub(super) fn parse_dollar_paren_block_pub(input: &str) -> PResult<'_, Expr> {
-    parse_dollar_paren_block(input)
-}
-
-fn parse_dollar_paren_block(input: &str) -> PResult<'_, Expr> {
-    use super::super::helpers::ws;
-    // Parse first statement
-    let (mut rest, _) = ws(input)?;
-    let mut stmts = Vec::new();
-    loop {
-        if rest.starts_with(')') || rest.is_empty() {
-            break;
-        }
-        let (r, stmt) = super::super::stmt::statement_pub(rest)?;
-        stmts.push(stmt);
-        let (r, _) = ws(r)?;
-        // Consume semicolons
-        let mut r = r;
-        while r.starts_with(';') {
-            r = &r[1..];
-            let (r2, _) = ws(r)?;
-            r = r2;
-        }
-        rest = r;
-    }
-    if stmts.len() < 2 {
-        return Err(PError::expected("statement block with multiple statements"));
-    }
-    let (rest, _) = parse_char(rest, ')')?;
-    Ok((
-        rest,
-        Expr::DoBlock {
-            body: stmts,
-            label: None,
-        },
-    ))
-}
-
-/// Detect Perl 5 special variables that are unsupported in Raku.
-/// Returns `Some(error_message)` if the input starts with a Perl 5 variable pattern,
-/// or `None` if it is a valid Raku variable.
-pub(crate) fn detect_perl5_scalar_var(input: &str) -> Option<String> {
-    // $^A through $^Z (uppercase) are Perl 5 variables.
-    // Raku placeholder parameters use lowercase: $^a through $^z.
-    if input.starts_with('^') && input.len() >= 2 {
-        let next = input.as_bytes()[1];
-        if next.is_ascii_uppercase() {
-            let var_char = next as char;
-            return Some(format!(
-                "X::Syntax::Perl5Var: Unsupported use of $^{} variable",
-                var_char
-            ));
-        }
-    }
-    // $" — Perl 5 list separator; in Raku use .join()
-    if input.starts_with('"') {
-        // Check that this is bare $" not inside a string interpolation context.
-        // $" followed by something that looks like assignment or end of statement.
-        // We need to be careful: in `"$foo"`, the `$` starts interpolation, not $".
-        // This check is called from scalar_var which is called in expression context,
-        // so $" at the start means the Perl 5 variable.
-        return Some(
-            "X::Syntax::Perl5Var: Unsupported use of $\" variable; in Raku please use .join() method"
-                .to_string(),
-        );
-    }
-    // $$ — Perl 5 PID; in Raku use $*PID
-    // $$ is Perl 5's PID variable only when followed by whitespace or assignment.
-    // In all other contexts, $$ is a valid Raku construct (e.g., $$x is scalar deref,
-    // $$$$ in a signature means four anonymous scalar params).
-    if let Some(after) = input.strip_prefix('$') {
-        let is_perl5_pid = after.is_empty()
-            || after.starts_with(' ')
-            || after.starts_with('\t')
-            || after.starts_with('\n')
-            || after.starts_with('\r')
-            || after.starts_with('=')
-            || after.starts_with(';');
-        if is_perl5_pid {
-            return Some(
-                "X::Syntax::Perl5Var: Unsupported use of $$ variable; in Raku please use $*PID"
-                    .to_string(),
-            );
-        }
-    }
-    // $' — Perl 5 postmatch; in Raku use $/.postmatch
-    // Only flag when followed by a space/operator, not when starting a string literal.
-    if let Some(after) = input.strip_prefix('\'') {
-        let looks_like_var = after.is_empty()
-            || after.starts_with(' ')
-            || after.starts_with('=')
-            || after.starts_with(';')
-            || after.starts_with('\n')
-            || after.starts_with(')')
-            || after.starts_with(',');
-        if looks_like_var {
-            return Some(
-                "X::Syntax::Perl5Var: Unsupported use of $' variable; in Raku please use $/.postmatch"
-                    .to_string(),
-            );
-        }
-    }
-    // $\ — Perl 5 output record separator; in Raku use .nl-out attribute
-    // Only flag when followed by whitespace+assignment, to avoid false positives
-    // in other contexts (e.g., `say $\` which should give "Confused").
-    if let Some(after) = input.strip_prefix('\\') {
-        let after_trimmed = after.trim_start();
-        let is_perl5_assignment = !after_trimmed.is_empty()
-            && after_trimmed.starts_with('=')
-            && !after_trimmed.starts_with("==")
-            && !after_trimmed.starts_with("=>");
-        if is_perl5_assignment {
-            return Some(
-                "X::Syntax::Perl5Var: Unsupported use of $\\ variable; in Raku please use the filehandle's .nl-out attribute"
-                    .to_string(),
-            );
-        }
-    }
-    // $& — Perl 5 match variable; in Raku use $/ or $<>.
-    // `$&foo` is a code-object deref (handled later), so only flag when the
-    // `&` is not the start of a routine name.
-    if let Some(after) = input.strip_prefix('&')
-        && !after.chars().next().is_some_and(is_raku_identifier_start)
-    {
-        return Some(
-            "X::Syntax::Perl5Var: Unsupported use of $& variable; in Raku please use $<>"
-                .to_string(),
-        );
-    }
-    // $` — Perl 5 prematch; in Raku use $/.prematch. A backtick never starts a
-    // valid Raku variable.
-    if input.starts_with('`') {
-        return Some(
-            "X::Syntax::Perl5Var: Unsupported use of $` variable; in Raku please use $/.prematch"
-                .to_string(),
-        );
-    }
-    // $| — Perl 5 autoflush; in Raku use the filehandle's .out-buffer attribute.
-    if input.starts_with('|') {
-        return Some(
-            "X::Syntax::Perl5Var: Unsupported use of $| variable; in Raku please use the filehandle's .out-buffer attribute"
-                .to_string(),
-        );
-    }
-    // $? — Perl 5 child error. `$?FILE`/`$?LINE`/`$?PACKAGE` etc. are valid Raku
-    // compile-time variables (a `?`-twigil before an identifier), so only flag a
-    // bare `$?`.
-    if let Some(after) = input.strip_prefix('?')
-        && !after.chars().next().is_some_and(is_raku_identifier_start)
-    {
-        return Some(
-            "X::Syntax::Perl5Var: Unsupported use of $? variable; in Raku please use $! for handling child errors also"
-                .to_string(),
-        );
-    }
-    // $@ — Perl 5 error variable; in Raku use $!. Only flag assignment position
-    // (`$@ = ...`): bare `$@` is allowed, and `$@!attr` is an item-context deref
-    // of a private array attribute.
-    if let Some(after) = input.strip_prefix('@') {
-        let trimmed = after.trim_start();
-        if trimmed.starts_with('=')
-            && !trimmed.starts_with("==")
-            && !trimmed.starts_with("=>")
-            && !trimmed.starts_with("=:=")
-        {
-            return Some(
-                "X::Syntax::Perl5Var: Unsupported use of $@ variable; in Raku please use $!"
-                    .to_string(),
-            );
-        }
-    }
-    // $; $, $. — Perl 5 separator/format variables. These characters are
-    // normally operators/separators (and `$.foo` is Raku attribute access), so
-    // only flag them in assignment position (`$; = ...`), which is how Perl 5
-    // code uses them and what distinguishes them from valid Raku constructs.
-    for (ch, suggestion) in [
-        (';', "real multidimensional hashes"),
-        (',', ".join() method"),
-        ('.', "the .kv method on e.g. .lines"),
-    ] {
-        if let Some(after) = input.strip_prefix(ch) {
-            let trimmed = after.trim_start();
-            let is_assignment = trimmed.starts_with('=')
-                && !trimmed.starts_with("==")
-                && !trimmed.starts_with("=>")
-                && !trimmed.starts_with("=:=");
-            if is_assignment {
-                return Some(format!(
-                    "X::Syntax::Perl5Var: Unsupported use of ${} variable; in Raku please use {}",
-                    ch, suggestion
-                ));
-            }
-        }
-    }
-    // $# followed by identifier — Perl 5 last index; in Raku use @array.end
-    if let Some(after) = input.strip_prefix('#')
-        && !after.is_empty()
-        && after.chars().next().is_some_and(is_raku_identifier_start)
-    {
-        let end = after
-            .find(|c: char| !is_raku_identifier_continue(c))
-            .unwrap_or(after.len());
-        let name = &after[..end];
-        return Some(format!(
-            "X::Syntax::Perl5Var: Unsupported use of $#{} variable; in Raku please use @{}.end",
-            name, name
-        ));
-    }
-    None
-}
-
-/// Detect Perl 5 special array/hash variables (`@-`, `@+`, `%-`, `%+`, `%!`)
-/// unsupported in Raku. `input` is the text after the sigil. Because `-`/`+`
-/// are operators and `%!foo`/`@!foo` are private attributes, these are only
-/// flagged in assignment position (`@- = ...`), matching how Perl 5 uses them.
-fn detect_perl5_sigil_var(sigil: char, input: &str) -> Option<String> {
-    let assignment_after = |after: &str| -> bool {
-        let trimmed = after.trim_start();
-        trimmed.starts_with('=')
-            && !trimmed.starts_with("==")
-            && !trimmed.starts_with("=>")
-            && !trimmed.starts_with("=:=")
-    };
-    // @-/@+/%-/%+ : match-position arrays/hashes; suggest .from / .to.
-    for (ch, suggestion) in [('-', ".from method"), ('+', ".to method")] {
-        if let Some(after) = input.strip_prefix(ch)
-            && assignment_after(after)
-        {
-            return Some(format!(
-                "X::Syntax::Perl5Var: Unsupported use of {}{} variable; in Raku please use {}",
-                sigil, ch, suggestion
-            ));
-        }
-    }
-    // %! : Perl 5 error hash. `%!foo` is a private attribute, so only flag the
-    // standalone variable in assignment position.
-    if sigil == '%'
-        && let Some(after) = input.strip_prefix('!')
-        && assignment_after(after)
-    {
-        return Some("X::Syntax::Perl5Var: Unsupported use of %! variable".to_string());
-    }
-    None
-}
+// ── pub(super): accessible from parser::primary (the parent of `var`)
+//    and all its descendants via `crate::parser::primary::var::` paths ──────
+pub(super) use scalar::{parse_dollar_paren_block_pub, parse_var_name_from_str, scalar_var};
+pub(super) use sigil_vars::{array_var, code_var, hash_var};
